Organise source files
