+++ tsc/usb/capability.ts
