+++ tsc/usb/descriptors.ts
