+++ tsc/usb/device.ts
