+++ tsc/usb/interface.ts
